backend/extract_pics_with_subscription/extract_pics_clone/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
Pillow==10.4.0

